tg_connect_bot/requirements.txt
aiogram==3.1.1
psycopg==3.1.12
python-decouple==3.8
beautifulsoup4~=4.12.2
aioimaplib~=1.0.1
openai==0.28.1
stripe~=6.7.0
selenium~=4.14.0
pydub~=0.25.1
playwright~=1.39.0
APScheduler~=3.10.4
requests~=2.31.0
bs4~=0.0.1
aiohttp~=3.8.5
webdriver-manager
google-auth
google-api-python-client
oauth2client